backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
pillow
PyJWT
python-dotenv
sqlparse
tzdata
resend
environs
whitenoise
psycopg2-binary
boto3
django-storages
gunicorn
dj-database-url
redis
hiredis
django-debug-toolbar